HIVE-2561 [jira] Add an annotation to UDFs to allow them to specify additional FILE/JAR resources necessary for execution
Summary:
Add requirements.

Often times UDFs will have dependencies to external JARs/FILEs.  It makes sense
for these to be encoded by the UDF (rather than having the caller remember the
set of files that need to be ADDed).  Let's add an annotation to UDFs which will
cause these resources to be auto-added.

Test Plan: EMPTY

Reviewers: JIRA, jsichi

Reviewed By: jsichi

CC: jsichi, jonchang, atfiore

Differential Revision: 507

git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1208075 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 531d3cc84291368f3c2c0950e23ce7aeeae7302f