HIVE-2767 [jira] Optionally use framed transport with metastore (Travis Crawford via Ashutosh Chauhan)
Summary:
Add support for optionally using the thrift framed transport, enabling
integration with environments where that is necessary.

Users may want/need to use thrift's framed transport when communicating with the
Hive MetaStore. This patch adds a new property
hive.metastore.thrift.framed.transport.enabled that enables the framed transport
(defaults to off, aka no change from before the patch). This property must be
set for both clients and the HMS server.

It wasn't immediately clear how to use the framed transport with SASL, so as
written an exception is thrown if you try starting the server with both options.
If SASL and the framed transport will indeed work together I can update the
patch (although I don't have a secured environment to test in).

Test Plan:
Tested locally that client and server can connect, both with and
without the flag. Tests pass.

Reviewers: JIRA, ashutoshc

Reviewed By: ashutoshc

Differential Revision: https://reviews.facebook.net/D2661

git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1325446 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: cad7e13a1aa76ffca762a8624032283190591e45